pyproject.toml
[project]
name = "task-queue-db"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil Pautov",email = "danpautoff@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
pre-commit = "^4.2.0"

[tool.poetry]
package-mode = false
